requirements.txt
Pillow==8.1.0
appdirs==1.4.4
autopep8==1.5.4
cycler==0.10.0
imageio==2.9.0
joblib==1.0.0
kiwisolver==1.3.1
llvmlite==0.35.0
matplotlib==3.3.3
meshio==4.3.8
numba==0.52.0
numpy-quaternion==2020.11.2.17.0.49
numpy==1.19.4
opencv-python==4.5.1.48
pandas==1.2.0
plyfile==0.7.2
pycodestyle==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.5
pyvista==0.27.4
scikit-learn==0.24.0
scipy==1.6.0
scooby==0.5.6
six==1.15.0
threadpoolctl==2.1.0
toml==0.10.2
vtk==9.0.1
